+++ handlers/route.go
@@ -1,7 +1,6 @@
 package handlers
 
 import (
-	"STUOJ/handlers/user"
 	"STUOJ/model"
 	"github.com/gin-gonic/gin"
 	"net/http"
@@ -40,10 +39,10 @@ func InitRoute() {
 func InitUserRoute() {
 	userRoute := ginServer.Group("/user")
 	{
-		userRoute.POST("/login", user.UserLogin)
-		userRoute.POST("/register", user.UserRegister)
-		userRoute.POST("/logout", user.UserLogout)
-		userRoute.POST("/data", user.UserData)
+		userRoute.POST("/login", UserLogin)
+		userRoute.POST("/register", UserRegister)
+		userRoute.POST("/logout", UserLogout)
+		userRoute.POST("/data", UserData)
 	}
 }
