--- handlers/user/data.go
@@ -1,6 +1,0 @@
-package user
-
-import "github.com/gin-gonic/gin"
-
-func UserData(c *gin.Context) {
-}
