--- handlers/user/login.go
@@ -1,12 +1,0 @@
-package user
-
-import "github.com/gin-gonic/gin"
-
-type ReqUserLogin struct {
-	Password string `json:"password" binding:"required"`
-	Email    string `json:"email" binding:"required"`
-}
-
-func UserLogin(c *gin.Context) {
-
-}
